Simple HTTP server. RIP uvicorns.
uvicorns proved to be incapable of allowing us to decode and handle mangled GET strings (that have the complete URL on the GET header). So we take the bruter-than-brute-force route of pulling apart the request ourselves with Python's built-in server. Where AI shines: completely rewriting something for a different framework. This time, it's all tested locally. But the real test is inside the container...